Head follows mouse
I don't know why but the head doesn't attach to the body when animated.

Head 'looks' at the location of the mouse now.

Would be nice if player faced the way the mouse is facing once the tongue is lashed out.

It just looks stupid when the player follows it at all times lol